+++ CHANGELOG.md
@@ -12,6 +12,17 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Fixes
 
+## v2.10.0 (2023-07-21)
+
+### New
+- added support for other partitions (`aws-cn` and `aws-us-gov`)
+- added CLI commands to be used as boilerplate code in deployspec for managing metadata
+  - ref: https://seed-farmer.readthedocs.io/en/latest/manifests.html#deployment-manifest
+### Changes
+- updated aws-codeseeder depenedency to 0.10.0
+
+### Fixes
+
 ## v2.9.4 (2023-06-29)
 
 ### New